キャラリストとシナリオリストのController
