+++ controller/main_menu.go
@@ -4,13 +4,12 @@ import (
 	"net/http"
 	"github.com/syo-sa1982/SeacherServerSide/model"
 	"log"
-	"html/template"
 	"encoding/json"
 )
 
 
 
-func (cntr Controller) MainIndex(c web.C, w http.ResponseWriter, r *http.Request) {
+func (cntr Controller) UserInfo(c web.C, w http.ResponseWriter, r *http.Request) {
 	var db = cntr.db
 	var UserInfo UserInfoAPI
 	var (
@@ -37,5 +36,29 @@ func (cntr Controller) MainIndex(c web.C, w http.ResponseWriter, r *http.Request
 
 	encoder := json.NewEncoder(w)
 	encoder.Encode(UserInfo)
+
+}
+
+func (cntr *Controller) PlayerList(c web.C, w http.ResponseWriter, r *http.Request) {
+	var db = cntr.db
+	user := model.User{}
+	job := model.JobMaster{}
+	playerBase := model.PlayerBase{}
+	playerStatus := model.PlayerStatus{}
+	r.ParseForm()
+	log.Println(r.FormValue("UUID"))
+
+	db.Model(user).Where("uuid = ?", r.FormValue("UUID")).Find(&user)
+	db.Model(playerStatus).Where("user_id = ?", user.ID).Find(&playerStatus)
+	db.Model(playerBase).Where("user_id = ?", user.ID).Find(&playerBase)
+	db.Model(job).Where("job_id = ?", playerStatus.JobID).Find(&job)
+
+	log.Println(playerStatus)
+
+	encoder := json.NewEncoder(w)
+	encoder.Encode(playerStatus)
+}
+
+func (cntr Controller) ScenarioList(c web.C, w http.ResponseWriter, r *http.Request) {
 
 }
